+++ apps/cookbook/src/app/showcase/list-showcase/list-showcase.component.ts
@@ -6,7 +6,9 @@ import { ListWithDividersExampleTemplate } from '~/app/examples/list-example/exa
 import { ListWithHeaderAndFooterExampleTemplate } from '~/app/examples/list-example/examples/header-and-footer';
 import { ListWithSectionsExampleTemplate } from '~/app/examples/list-example/examples/sections';
 import { ListWithSectionsAndColoredItemsExampleTemplate } from '~/app/examples/list-example/examples/sections-and-colored-items';
+import { ListWithSectionsAndStandAloneExampleTemplate } from '~/app/examples/list-example/examples/sections-with-standalone';
 import { ListSelectableItemsExampleTemplate } from '~/app/examples/list-example/examples/selectable-items';
+import { ListWithStandAloneExampleTemplate } from '~/app/examples/list-example/examples/stand-alone';
 
 @Component({
   selector: 'cookbook-list-showcase',
@@ -20,6 +22,8 @@ export class ListShowcaseComponent {
   sectionsAndColoredItemsExampleTemplate: string = ListWithSectionsAndColoredItemsExampleTemplate;
   headerAndFooterExampleTemplate: string = ListWithHeaderAndFooterExampleTemplate;
   sectionsExampleTemplate: string = ListWithSectionsExampleTemplate;
+  standAloneExampleTemplate: string = ListWithStandAloneExampleTemplate;
+  sectionsAndStandAloneExampleTemplate: string = ListWithSectionsAndStandAloneExampleTemplate;
   items: string = ListItemsExampleTemplate;
 
   exampleHtml = exampleHtml;