+++ apps/flows/src/app/transaction-details-flow/transaction-details-flow.component.ts
@@ -1,0 +1,11 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'kirbydesign-transaction-details-flow',
+  templateUrl: './transaction-details-flow.component.html',
+})
+export class TransactionDetailsFlowComponent implements OnInit {
+  constructor() {}
+
+  ngOnInit(): void {}
+}