+++ apps/flows/src/app/transaction.ts
@@ -1,0 +1,20 @@
+export interface Transaction {
+  id: string;
+  text: string;
+  amount: {
+    amount: number;
+    currencyCode: string;
+  };
+  iconName: string;
+  date: {
+    epoch: number;
+    utc: string;
+  };
+  category: {
+    id: string;
+  };
+  balance: {
+    amount: number;
+  };
+  balanced: boolean;
+}