+++ libs/designsystem/src/lib/components/index.ts
@@ -15,6 +15,8 @@ export * from './item-sliding';
 export { CheckboxComponent } from './checkbox/checkbox.component';
 export { ChipComponent } from './chip/chip.component';
 
+export * from './data-table';
+
 export { DividerComponent } from './divider/divider.component';
 
 export { DropdownComponent } from './dropdown/dropdown.component';
@@ -63,3 +65,5 @@ export { ToggleComponent } from './toggle/toggle.component';
 
 export { ProgressCircleComponent } from './progress-circle/progress-circle.component';
 export * from './toggle-button';
+
+export * from './page-local-navigation';