+++ libs/designsystem/src/lib/components/list/index.ts
@@ -1,5 +1,5 @@
 export { ListComponent } from './list.component';
-export { ListShape } from './list.component';
+export { ListShape, StandAloneSpacing } from './list.component';
 export { ListHeaderComponent } from './list-header/list-header.component';
 export { ListSectionHeaderComponent } from './list-section-header/list-section-header.component';
 export { ListItemComponent } from './list-item/list-item.component';
@@ -17,5 +17,3 @@ export { ListSectionHeaderDirective } from './list.directive';
 export { ListFooterDirective } from './list.directive';
 
 export { ListExperimentalComponent } from './list-experimental/list-experimental.component';
-
-export { GroupByPipe } from './pipes/group-by.pipe';