+++ libs/designsystem/src/lib/components/page/page.component.spec.ts
@@ -11,7 +11,7 @@ import { TestHelper } from '../../testing/test-helper';
 import { WindowRef } from '../../types/window-ref';
 import { ButtonComponent } from '../button/button.component';
 import { ModalNavigationService } from '../modal/services/modal-navigation.service';
-import { TabsComponent } from '../tabs';
+import { selectedTabClickEvent, TabsComponent } from '../tabs';
 
 const { size, fontWeight } = DesignTokenHelper;
 
@@ -442,6 +442,17 @@ describe('PageComponent', () => {
 
     it('should not be available when "refresh" is not subscribed to', () => {
       expect(spectator.query(IonRefresher)).toBeNull();
+    });
+  });
+
+  describe('tab navigation', () => {
+    it('should scroll to top when tab is clicked', () => {
+      const scrollToTopSpy = jasmine.createSpy();
+      (spectator.component as any).content.scrollToTop = scrollToTopSpy;
+
+      window.dispatchEvent(new Event(selectedTabClickEvent));
+
+      expect(scrollToTopSpy).toHaveBeenCalledTimes(1);
     });
   });